requirements.txt
streamlit
python-dotenv
agents
openai 
httpx
